bugfix: was getting the wrong docstring for method_dispatch resources
Problem occurred when the decorated function had not docstring, which
means I was using "doc = None" and passing in the method_dispatcher
object to the regsitry. When that happens, the registry tries to get
the docstring from the handler, which was the method_dispacther
instance, which is an internal class that contains a docstring, but
one which is irrelavant to the registry. Fix it by using "" if the
decorated function does not have a docstring