Fixed failure where I was using argv incorrectly.
lib/commandline.py was not using the argv passed in to it. I had fixed
that to make test_commandline.py work, but I did it in such a way that
the command-line scripts no longer worked. Fixing that identified a
bug in how my new test_log_rotate worked.

Should be all better now.